backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.11.0
python-multipart==0.0.9
crewai==0.11.2
langchain==0.1.0
langchain-openai==0.0.5
langchain-groq==0.1.5
langchain-google-genai>=0.1.5
pymupdf==1.23.8
python-docx==1.1.0
beautifulsoup4==4.12.2
nltk==3.8.1
spacy==3.7.2
numpy==1.26.3
pandas==2.1.4
python-dotenv==1.0.0
aiofiles==23.2.1
pytest==7.4.3
httpx==0.26.0
